+++ client/.env.example
@@ -1,0 +1,2 @@
+REACT_APP_AUTH_DOMAIN="YOUR AUTH0 DOMAIN"
+REACT_APP_AUTH_CLIENTID="YOUR AUTH0 CLIENT ID"
