+++ client/.gitignore
@@ -2,3 +2,5 @@ node_modules
 dist
 package-lock.json
 .cache
+.parcel-cache
+.env
