+++ client/package.json
@@ -9,7 +9,9 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@auth0/auth0-react": "^1.3.0",
 		"app": "^0.1.0",
+		"parcel": "^2.0.0-beta.2",
 		"react": "^16.14.0",
 		"react-dom": "^17.0.1",
 		"react-motion": "^0.5.2",
