+++ client/src/components/componentiChiamata.js
@@ -1,5 +1,5 @@
 import React from "react";
-import api from "./api";
+import api from "../api";
 
 export function BarraChiamata(props) {
 	let ordine = [
